Add assetloader, modeldata, and mesh classes - Added AssetLoader to load a file using Assimp - Added ModelData to traverse the scene tree and create instances of   meshdata that copy the aiMesh data into vertex and index vectors. - Added Mesh to construct vertex and index buffers from meshdata - Changed vertex and index buffers to take const vector references   instead of pointers

+++ Graphics-Demo/src/d3d/inputlayout.cpp
@@ -17,7 +17,7 @@ namespace d3d
 	{
 		DB_LOG("Creating InputLayout, inputDescs size: " << inputDescs.size() << ", bytcode size: " << bytecode.size() << '\n');
 
-		HR(app.getDevice().CreateInputLayout(inputDescs.data(), inputDescs.size(), bytecode.data(), bytecode.size(), &m_inputLayout));
+		HR(app.getDevice().CreateInputLayout(inputDescs.data(), static_cast<UINT>(inputDescs.size()), bytecode.data(), bytecode.size(), &m_inputLayout));
 
 		assert(m_inputLayout.Get());
 	}
